refac: fileservice/folder.rs
