+++ apps/desktop/src-tauri/src/services/file_service/folder/import.rs
@@ -1,16 +1,16 @@
-use anyhow::{anyhow, bail, Context, Result};
-use async_recursion::async_recursion;
-use serde::Serialize;
-use sqlx::{Sqlite, Transaction};
 use std::{
-    collections::{BTreeSet, HashMap, HashSet},
+    collections::HashSet,
     path::{Path, PathBuf},
     sync::Arc,
-    time::{Duration, Instant},
+    time::Instant,
 };
-use tauri::{AppHandle, Emitter};
+
+use anyhow::{anyhow, bail, Context, Result};
+use async_recursion::async_recursion;
+use sqlx::{Sqlite, Transaction};
+use tauri::AppHandle;
 use tokio::{fs, sync::Mutex};
-use tracing::{info, warn};
+use tracing::warn;
 
 use crate::{
     db::repository::{
@@ -21,577 +21,28 @@ use crate::{
     models::{
         file::{
             FileInfo, FileType, FolderData, FolderOptionUpdatePayload,
-            FolderPreview, FolderPreviewFileStat, FolderPreviewNode,
-            FolderPreviewSummary, FolderSelection, RealFolderData,
+            FolderSelection, RealFolderData,
         },
         node::Node,
     },
     services::{
         db::DB_MANAGER,
+        file_service::job_queue::{enqueue_base_job, BaseThumbnailJob},
         storage_root::{self, ensure_storage_root_and_real_folder},
     },
     utils::{file_utils::file_mtime_epoch, path_utils::normalize_path},
 };
 
 use super::{
-    job_queue::{enqueue_base_job, BaseThumbnailJob},
-    utils::check_is_hidden,
+    metrics::UpsertFolderMetrics,
+    progress::ImportProgressTracker,
+    selection::{
+        derive_root_extension_list, normalize_extension_list,
+        relative_path_key, ExtensionFilter, SelectionPlan,
+    },
 };
 
-#[derive(Default)]
-struct UpsertFolderMetrics {
-    total_elapsed: Duration,
-    folder_creation: Duration,
-    folder_creation_count: u64,
-    tree_scanning: Duration,
-    tree_scanning_count: u64,
-    upsert_file: Duration,
-    upsert_file_count: u64,
-    file_size_buckets: HashMap<FileSizeBucket, BucketStats>,
-}
-
-impl UpsertFolderMetrics {
-    fn record_folder_creation(&mut self, elapsed: Duration) {
-        self.folder_creation += elapsed;
-        self.folder_creation_count += 1;
-    }
-
-    fn record_tree_scanning(&mut self, elapsed: Duration) {
-        self.tree_scanning += elapsed;
-        self.tree_scanning_count += 1;
-    }
-
-    fn record_upsert_file(&mut self, size: Option<i64>, elapsed: Duration) {
-        self.upsert_file += elapsed;
-        self.upsert_file_count += 1;
-
-        let bucket = FileSizeBucket::from_size(size);
-        let entry = self.file_size_buckets.entry(bucket).or_default();
-        entry.duration += elapsed;
-        entry.count += 1;
-    }
-
-    fn log(&self, root_dir: &Path) {
-        let mut bucket_entries: Vec<_> =
-            self.file_size_buckets.iter().collect();
-        bucket_entries.sort_by_key(|(bucket, _)| bucket.order());
-        let bucket_summary = bucket_entries
-            .into_iter()
-            .map(|(bucket, stats)| {
-                format!(
-                    "{}: {}ms/{}",
-                    bucket.label(),
-                    stats.duration.as_millis(),
-                    stats.count
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(", ");
-        let bucket_summary = if bucket_summary.is_empty() {
-            "none".to_string()
-        } else {
-            bucket_summary
-        };
-
-        info!(
-            "Upsert Folder Metrics
-                dir                : {}
-                total_ms           : {}
-                tree_scan          : {} ms ({} runs)
-                folder_creation    : {} ms ({} runs)
-                upsert_file        : {} ms ({} runs)
-                file_size_buckets  : {}
-            ",
-            root_dir.display(),
-            self.total_elapsed.as_millis(),
-            self.tree_scanning.as_millis(),
-            self.tree_scanning_count,
-            self.folder_creation.as_millis(),
-            self.folder_creation_count,
-            self.upsert_file.as_millis(),
-            self.upsert_file_count,
-            bucket_summary,
-        );
-    }
-}
-
-#[derive(Clone)]
-struct FolderUpsertConfig<'a> {
-    recursive: bool,
-    make_virtual_folder: Option<bool>,
-    selection: Option<&'a SelectionPlan>,
-    root_path: &'a Path,
-    extension_filter: &'a ExtensionFilter,
-    progress: Option<Arc<Mutex<ImportProgressTracker>>>,
-}
-
-impl<'a> FolderUpsertConfig<'a> {
-    fn with_make_virtual_folder(
-        mut self,
-        make_virtual_folder: Option<bool>,
-    ) -> Self {
-        self.make_virtual_folder = make_virtual_folder;
-        self
-    }
-}
-
-struct FileEntryParams<'a> {
-    parent_real_folder_id: &'a str,
-    parent_virtual_folder_id: &'a str,
-    entry: &'a fs::DirEntry,
-    allowed_types: Option<&'a HashSet<FileType>>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-enum ImportState {
-    Running,
-    Completed,
-    Failed,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct FolderImportProgressPayload {
-    folder_id: String,
-    state: ImportState,
-    processed_bytes: u64,
-    total_bytes: Option<u64>,
-    processed_files: u64,
-    total_files: Option<u64>,
-    elapsed_ms: u64,
-}
-
-struct ImportProgressTracker {
-    app_handle: AppHandle,
-    moa_id: String,
-    folder_id: String,
-    total_bytes: Option<u64>,
-    total_files: Option<u64>,
-    processed_bytes: u64,
-    processed_files: u64,
-    started_at: Instant,
-}
-
-impl ImportProgressTracker {
-    fn new(
-        app_handle: AppHandle,
-        moa_id: String,
-        folder_id: String,
-        total_bytes: Option<u64>,
-        total_files: Option<u64>,
-    ) -> Self {
-        Self {
-            app_handle,
-            moa_id,
-            folder_id,
-            total_bytes,
-            total_files,
-            processed_bytes: 0,
-            processed_files: 0,
-            started_at: Instant::now(),
-        }
-    }
-
-    fn emit(&self, state: ImportState) {
-        let elapsed = self.started_at.elapsed().as_millis();
-        let elapsed_ms = if elapsed > u128::from(u64::MAX) {
-            u64::MAX
-        } else {
-            elapsed as u64
-        };
-
-        let payload = FolderImportProgressPayload {
-            folder_id: self.folder_id.clone(),
-            state,
-            processed_bytes: self.processed_bytes,
-            total_bytes: self.total_bytes,
-            processed_files: self.processed_files,
-            total_files: self.total_files,
-            elapsed_ms,
-        };
-
-        let topic = format!("folder-import://progress/{}", self.moa_id);
-        let _ = self.app_handle.emit(&topic, payload);
-    }
-
-    fn notify_start(&self) {
-        self.emit(ImportState::Running);
-    }
-
-    fn record_file(&mut self, bytes: Option<i64>) {
-        self.processed_files = self.processed_files.saturating_add(1);
-        if let Some(size) = bytes {
-            if size > 0 {
-                self.processed_bytes =
-                    self.processed_bytes.saturating_add(size as u64);
-            }
-        }
-        self.emit(ImportState::Running);
-    }
-
-    fn finish(&mut self) {
-        if let Some(total) = self.total_bytes {
-            if self.processed_bytes < total {
-                self.processed_bytes = total;
-            }
-        } else {
-            self.total_bytes = Some(self.processed_bytes);
-        }
-
-        if let Some(total) = self.total_files {
-            if self.processed_files < total {
-                self.processed_files = total;
-            }
-        } else {
-            self.total_files = Some(self.processed_files);
-        }
-
-        self.emit(ImportState::Completed);
-    }
-
-    fn fail(&self) {
-        self.emit(ImportState::Failed);
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-enum FileSizeBucket {
-    Unknown,
-    UpTo1Mb,
-    OneToTenMb,
-    TenToHundredMb,
-    OverHundredMb,
-}
-
-impl FileSizeBucket {
-    fn from_size(size: Option<i64>) -> Self {
-        match size {
-            None => FileSizeBucket::Unknown,
-            Some(bytes) if bytes < 1_000_000 => FileSizeBucket::UpTo1Mb,
-            Some(bytes) if bytes < 10_000_000 => FileSizeBucket::OneToTenMb,
-            Some(bytes) if bytes < 100_000_000 => {
-                FileSizeBucket::TenToHundredMb
-            }
-            Some(_) => FileSizeBucket::OverHundredMb,
-        }
-    }
-
-    fn label(self) -> &'static str {
-        match self {
-            FileSizeBucket::Unknown => "unknown",
-            FileSizeBucket::UpTo1Mb => "<1MB",
-            FileSizeBucket::OneToTenMb => "1-10MB",
-            FileSizeBucket::TenToHundredMb => "10-100MB",
-            FileSizeBucket::OverHundredMb => ">=100MB",
-        }
-    }
-
-    fn order(self) -> u8 {
-        match self {
-            FileSizeBucket::Unknown => 0,
-            FileSizeBucket::UpTo1Mb => 1,
-            FileSizeBucket::OneToTenMb => 2,
-            FileSizeBucket::TenToHundredMb => 3,
-            FileSizeBucket::OverHundredMb => 4,
-        }
-    }
-}
-
-#[derive(Default)]
-struct BucketStats {
-    duration: Duration,
-    count: u64,
-}
-
-const FILE_TYPE_ORDER: [FileType; 7] = [
-    FileType::Image,
-    FileType::Video,
-    FileType::Document,
-    FileType::GraphicTool,
-    FileType::Audio,
-    FileType::Archive,
-    FileType::Unknown,
-];
-
-#[derive(Default)]
-struct SelectionPlan {
-    entries: HashMap<String, SelectionNode>,
-}
-
-#[derive(Clone, Default)]
-struct SelectionNode {
-    include: bool,
-    allowed_types: Option<HashSet<FileType>>,
-}
-
-impl SelectionPlan {
-    fn from_selection(selection: FolderSelection) -> Self {
-        let mut entries = HashMap::new();
-        for entry in selection.entries {
-            let key = normalize_relative_key(&entry.relative_path);
-            let allowed_types = entry
-                .file_types
-                .map(|types| types.into_iter().collect::<HashSet<FileType>>());
-            entries.insert(
-                key,
-                SelectionNode { include: entry.include, allowed_types },
-            );
-        }
-
-        SelectionPlan { entries }
-    }
-
-    fn get(&self, key: &str) -> Option<&SelectionNode> {
-        self.entries.get(key)
-    }
-}
-
-fn normalize_extension_list(list: Vec<String>) -> Vec<String> {
-    let mut seen: HashSet<String> = HashSet::new();
-    let mut normalized = Vec::new();
-
-    for value in list {
-        let trimmed = value.trim().trim_start_matches('.').to_lowercase();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if seen.insert(trimmed.clone()) {
-            normalized.push(trimmed);
-        }
-    }
-
-    normalized
-}
-
-fn derive_root_extension_list(
-    selection: Option<&SelectionPlan>,
-) -> Option<Vec<String>> {
-    let root = selection?.get("");
-    let allowed = root?.allowed_types.as_ref()?;
-
-    let mut set: BTreeSet<String> = BTreeSet::new();
-    for file_type in allowed.iter() {
-        for ext in file_type.extensions() {
-            set.insert(ext.to_string());
-        }
-    }
-
-    if set.is_empty() {
-        None
-    } else {
-        Some(set.into_iter().collect())
-    }
-}
-
-#[derive(Default, Clone)]
-pub struct ExtensionFilter {
-    pub include: Option<HashSet<String>>,
-    pub exclude: HashSet<String>,
-}
-
-impl ExtensionFilter {
-    pub fn new(include: &[String], exclude: &[String]) -> Self {
-        let include = if include.is_empty() {
-            None
-        } else {
-            Some(include.iter().cloned().collect())
-        };
-
-        let exclude = exclude.iter().cloned().collect();
-
-        ExtensionFilter { include, exclude }
-    }
-
-    pub fn allows(&self, path: &Path) -> bool {
-        let ext = path
-            .extension()
-            .and_then(|value| value.to_str())
-            .map(|value| value.trim_start_matches('.').to_lowercase());
-
-        if let Some(ref ext) = ext {
-            if self.exclude.contains(ext) {
-                return false;
-            }
-
-            if let Some(include) = &self.include {
-                return include.contains(ext);
-            }
-        } else if self.include.is_some() {
-            return false;
-        }
-
-        true
-    }
-}
-
-#[derive(Default)]
-struct PreviewAccumulator {
-    total_folders: u64,
-    total_files: u64,
-    total_bytes: u64,
-    file_type_totals: HashMap<FileType, PreviewFileStats>,
-}
-
-#[derive(Default)]
-struct PreviewFileStats {
-    count: u64,
-    bytes: u64,
-}
-
-fn normalize_relative_key(value: &str) -> String {
-    if value.is_empty() {
-        return String::new();
-    }
-
-    let replaced = value.replace('\\', "/");
-    let trimmed = replaced.trim_matches('/');
-    if trimmed.is_empty() || trimmed == "." {
-        String::new()
-    } else {
-        trimmed.to_string()
-    }
-}
-
-fn relative_path_key(root: &Path, current: &Path) -> String {
-    if let Ok(relative) = current.strip_prefix(root) {
-        if relative.as_os_str().is_empty() {
-            return String::new();
-        }
-        return join_path_components(relative);
-    }
-
-    join_path_components(current)
-}
-
-fn join_path_components(path: &Path) -> String {
-    path.components()
-        .map(|c| c.as_os_str().to_string_lossy())
-        .collect::<Vec<_>>()
-        .join("/")
-}
-
-fn stats_map_to_vec(
-    map: &HashMap<FileType, PreviewFileStats>,
-) -> Vec<FolderPreviewFileStat> {
-    let mut out = Vec::new();
-    for file_type in FILE_TYPE_ORDER {
-        if let Some(stats) = map.get(&file_type) {
-            if stats.count > 0 {
-                out.push(FolderPreviewFileStat {
-                    file_type,
-                    count: stats.count,
-                    bytes: stats.bytes,
-                });
-            }
-        }
-    }
-
-    out
-}
-
-/// Traverse the filesystem and produce a preview tree for the selected folder.
-pub async fn collect_folder_preview(path: &Path) -> Result<FolderPreview> {
-    let norm = normalize_path(path);
-
-    let mut accumulator = PreviewAccumulator::default();
-    let root =
-        collect_folder_preview_impl(&norm, &norm, &mut accumulator).await?;
-
-    let summary = FolderPreviewSummary {
-        total_folders: accumulator.total_folders,
-        total_files: accumulator.total_files,
-        total_bytes: accumulator.total_bytes,
-        file_type_totals: stats_map_to_vec(&accumulator.file_type_totals),
-    };
-
-    Ok(FolderPreview { root, summary })
-}
-
-#[async_recursion]
-async fn collect_folder_preview_impl(
-    abs_dir: &Path,
-    root: &Path,
-    accumulator: &mut PreviewAccumulator,
-) -> Result<FolderPreviewNode> {
-    accumulator.total_folders += 1;
-
-    let mut dir = fs::read_dir(abs_dir)
-        .await
-        .with_context(|| format!("failed to read_dir {:?}", abs_dir))?;
-
-    let mut children = Vec::new();
-    let mut local_stats: HashMap<FileType, PreviewFileStats> = HashMap::new();
-    let mut total_files = 0_u64;
-    let mut total_bytes = 0_u64;
-
-    while let Some(entry) = dir
-        .next_entry()
-        .await
-        .with_context(|| format!("failed to read entry under {:?}", abs_dir))?
-    {
-        let entry_path = entry.path();
-
-        if check_is_hidden(&entry_path) {
-            continue;
-        }
-
-        let file_type = entry.file_type().await.with_context(|| {
-            format!("failed to get file_type for {:?}", entry_path)
-        })?;
-
-        if file_type.is_symlink() {
-            continue;
-        }
-
-        if file_type.is_dir() {
-            let child =
-                collect_folder_preview_impl(&entry_path, root, accumulator)
-                    .await?;
-            children.push(child);
-            continue;
-        }
-
-        if file_type.is_file() {
-            let kind = FileType::from(entry_path.as_path());
-            let metadata = entry.metadata().await.with_context(|| {
-                format!("failed to get metadata for {:?}", entry_path)
-            })?;
-            let size = metadata.len();
-
-            total_files += 1;
-            total_bytes += size;
-
-            let local_entry = local_stats.entry(kind).or_default();
-            local_entry.count += 1;
-            local_entry.bytes += size;
-
-            let global_entry =
-                accumulator.file_type_totals.entry(kind).or_default();
-            global_entry.count += 1;
-            global_entry.bytes += size;
-
-            accumulator.total_files += 1;
-            accumulator.total_bytes += size;
-        }
-    }
-
-    let name = abs_dir
-        .file_name()
-        .and_then(|n| n.to_str())
-        .map(|s| s.to_string())
-        .unwrap_or_else(|| abs_dir.to_string_lossy().into_owned());
-
-    Ok(FolderPreviewNode {
-        name,
-        path: abs_dir.to_string_lossy().into_owned(),
-        relative_path: relative_path_key(root, abs_dir),
-        total_files,
-        total_bytes,
-        file_stats: stats_map_to_vec(&local_stats),
-        children,
-    })
-}
+use crate::services::file_service::utils::check_is_hidden;
 
 /// Create a new virtual folder inside a transaction.
 pub async fn create_folder(moa_id: String, data: FolderData) -> Result<Node> {
@@ -611,6 +62,7 @@ pub async fn create_folder(moa_id: String, data: FolderData) -> Result<Node> {
 }
 
 /// Mount the first physical folder selected by the user into the virtual tree.
+#[allow(clippy::too_many_arguments)]
 pub async fn first_mount_folder(
     app: AppHandle,
     moa_id: String,
@@ -851,11 +303,38 @@ pub async fn update_virtual_folder_options(
 }
 
 /// Spawn a background job that scans a real folder.
-pub async fn start_scan_job(
+pub(crate) async fn start_scan_job(
     _moa_id: String,
     _real_folder_id: String,
 ) -> Result<String> {
     Ok(String::new())
+}
+
+#[derive(Clone)]
+struct FolderUpsertConfig<'a> {
+    recursive: bool,
+    make_virtual_folder: Option<bool>,
+    selection: Option<&'a SelectionPlan>,
+    root_path: &'a Path,
+    extension_filter: &'a ExtensionFilter,
+    progress: Option<Arc<Mutex<ImportProgressTracker>>>,
+}
+
+impl<'a> FolderUpsertConfig<'a> {
+    fn with_make_virtual_folder(
+        mut self,
+        make_virtual_folder: Option<bool>,
+    ) -> Self {
+        self.make_virtual_folder = make_virtual_folder;
+        self
+    }
+}
+
+struct FileEntryParams<'a> {
+    parent_real_folder_id: &'a str,
+    parent_virtual_folder_id: &'a str,
+    entry: &'a fs::DirEntry,
+    allowed_types: Option<&'a HashSet<FileType>>,
 }
 
 /// Recursively upsert folder and file information.
