+++ apps/desktop/src-tauri/src/services/file_service/folder/mod.rs
@@ -1,0 +1,11 @@
+mod import;
+mod metrics;
+mod preview;
+mod progress;
+mod selection;
+
+pub use import::{
+    create_folder, ensure_real_folder, fetch_one_file_path, first_mount_folder,
+    sync_virtual_folder, update_virtual_folder_options,
+};
+pub use preview::collect_folder_preview;
